+++ src/main/java/fipe/preco/preco_fipe/repository/UserRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
 
-    Optional<User> findByEmailAndIdNot(String email, Integer id);
+    Optional<User> findByEmailAndIdNot(String email, Long id);
 }
